requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
pydantic==1.10.8
bcrypt==4.0.1
python-jose==3.3.0
passlib==1.7.4
mongoengine==0.27.0
pytest==7.4.0
httpx==0.24.1